fix: move helper outside of pages dir
